+++ src/components/screens/forms/form-screen.tsx
@@ -12,6 +12,7 @@ import Mail from "mail/mail";
 import ConfirmDialog from "components/generic/confirm-dialog";
 import EmailDialog from "components/generic/email-dialog";
 import Form from "components/generic/form";
+import ReplySaved from "./form/ReplySaved";
 
 /**
  * Component props
@@ -410,61 +411,6 @@ const FormScreen: React.FC<Props> = () => {
   };
 
   /**
-   * Renders reply saved dialog
-   */
-  const renderReplySaved = () => {
-    const replyEditLink = getReplyEditLink();
-
-    if (replyEditLink) {
-      return (
-        <Snackbar open={ replySavedVisible } onClose={ () => setReplySavedVisible(false) } anchorOrigin={{ vertical: "top", horizontal: "center" }}>
-          <Alert onClose={ () => setReplySavedVisible(false) } severity="success">
-            <span>
-              {" "}
-              { strings.formScreen.replySaved }
-              {" "}
-            </span>
-            <p>
-              {" "}
-              { strings.formScreen.replyEdit }
-              {" "}
-            </p>
-            <div style={{
-              textOverflow: "ellipsis", overflow: "hidden", maxWidth: "50vw", whiteSpace: "nowrap"
-            }}
-            >
-              <a href={ replyEditLink }>
-                {" "}
-                { replyEditLink }
-                {" "}
-              </a>
-            </div>
-            <p>
-              { strings.formScreen.replyEditEmailText }
-              <Link href="#" onClick={ onReplyEmailLinkClick }>
-                {" "}
-                { strings.formScreen.replyEditEmailLink }
-                {" "}
-              </Link>
-            </p>
-          </Alert>
-        </Snackbar>
-      );
-    }
-    return (
-      <Snackbar open={ replySavedVisible } onClose={ () => setReplySavedVisible(false) } anchorOrigin={{ vertical: "top", horizontal: "center" }}>
-        <Alert onClose={ () => setReplySavedVisible(false) } severity="success">
-          <span>
-            {" "}
-            { strings.formScreen.replySaved }
-            {" "}
-          </span>
-        </Alert>
-      </Snackbar>
-    );
-  };
-
-  /**
    * Renders reply email dialog
    */
   const renderReplyEmailDialog = () => {
@@ -539,7 +485,12 @@ const FormScreen: React.FC<Props> = () => {
     <BasicLayout>
       <div>
         { renderForm() }
-        { renderReplySaved() }
+        <ReplySaved
+          getReplyEditLink={ getReplyEditLink }
+          replySavedVisible={ replySavedVisible }
+          onReplyEmailLinkClick={ onReplyEmailLinkClick }
+          setReplySavedVisible={ setReplySavedVisible }
+        />
         { renderReplyEmailDialog() }
         { renderReplyDelete() }
         { renderReplyDeleteConfirm() }
